Clean up tests; add notebook demo
